+++ allennlp/models/encoder_decoders/simple_seq2seq.py
@@ -442,7 +442,7 @@ class SimpleSeq2Seq(Model):
         # Ensure mask is also a FloatTensor. Or else the multiplication within
         # attention will complain.
         # shape: (batch_size, max_input_sequence_length, encoder_output_dim)
-        encoder_outputs_mask = encoder_outputs_mask.float()
+        encoder_outputs_mask = encoder_outputs_mask.type_as(encoder_outputs)
 
         # shape: (batch_size, max_input_sequence_length)
         input_weights = self._attention(
